Adjust the code to remove any duplication created in previous commit
