+++ Sig.App.Backend/Gql/Schema/GraphTypes/MarketGroupGraphType.cs
@@ -2,7 +2,6 @@
 using GraphQL.DataLoader;
 using MediatR;
 using NodaTime.Extensions;
-using Sig.App.Backend.DbModel.Entities.CashRegisters;
 using Sig.App.Backend.DbModel.Entities.MarketGroups;
 using Sig.App.Backend.Extensions;
 using Sig.App.Backend.Gql.Bases;
@@ -103,38 +102,7 @@ namespace Sig.App.Backend.Gql.Schema.GraphTypes
             var startInstant = startDate.ToInstant();
             var endInstant = endDate.ToInstant();
 
-            return transactions.Where(x => IsTransactionBetweenDate(x, startInstant, endInstant) && IsTransactionInCashRegister(x, cashRegisters));
-        }
-
-        private bool IsTransactionBetweenDate(ITransactionGraphType x, NodaTime.Instant startInstant, NodaTime.Instant endInstant)
-        {
-            var createdAtInstant = x.CreatedAt().ToInstant();
-
-            return startInstant <= createdAtInstant && createdAtInstant < endInstant;
-        }
-
-        private bool IsTransactionInCashRegister(ITransactionGraphType x, Id[] cashRegisters)
-        {
-            if (cashRegisters.Length > 0)
-            {
-                var cashRegisterIds = cashRegisters.Select(x => x.LongIdentifierForType<CashRegister>());
-                if (x is RefundTransactionGraphType rtgt)
-                {
-                    if (cashRegisterIds.Any(id => id == rtgt.CashRegisterId))
-                    {
-                        return true;
-                    }
-                }
-                if (x is PaymentTransactionGraphType ptgt)
-                {
-                    if (cashRegisterIds.Any(id => id == ptgt.CashRegisterId))
-                    {
-                        return true;
-                    }
-                }
-                return false;
-            }
-            return true;
+            return transactions.Where(x => TransactionGraphTypeHelper.IsTransactionBetweenDate(x, startInstant, endInstant) && TransactionGraphTypeHelper.IsTransactionInCashRegister(x, cashRegisters));
         }
     }
 }
